# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Ynov API for YnovImmo"
  version: "1.0.0"
  title: "Ynov API pour l'application YnovImmo"
host: "localhost:8080"
basePath: "/api/v1"
schemes:
- "http"
paths:
...
  /real-estate-visit:
    get:
      tags:
      - "real-estate-visit"
      summary: "get all real-estate-visit by pagination"
      description: ""
      operationId: "real_estate_visitGET"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where\
          \ id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
        collectionFormat: "multi"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "limit returning object count"
        required: false
        type: "integer"
        default: 2
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/RealEstateVisitPagination"
      x-swagger-router-controller: "RealEstateVisit"
    post:
      tags:
      - "real-estate-visit"
      summary: "create real-estate-visit"
      description: "create real-estate-visit"
      operationId: "real_estate_visitPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate-visit"
        required: true
        schema:
          $ref: "#/definitions/RealEstateVisit"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "RealEstateVisit"
    patch:
      tags:
      - "real-estate-visit"
      summary: "update real-estate-visit"
      description: "update real-estate-visit"
      operationId: "real_estate_visitPATCH"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate-visit"
        required: true
        schema:
          $ref: "#/definitions/RealEstateVisit"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "RealEstateVisit"
  /real-estate-visit/{Id}:
    get:
      tags:
      - "real-estate-visit"
      summary: "get a real-estate-visit by ID"
      description: "get a real-estate-visit by ID"
      operationId: "real_estate_visitIdGET"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate-visit to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/RealEstateVisit"
      x-swagger-router-controller: "RealEstateVisit"
    delete:
      tags:
      - "real-estate-visit"
      summary: "Destroy a real-estate-visit by ID"
      description: "delete a real-estate-visit by ID"
      operationId: "real_estate_visitIdDELETE"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate-visit to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "RealEstateVisit"
...
  /user:
    get:
      tags:
      - "user"
      summary: "get all user by pagination"
      description: ""
      operationId: "userGET"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where\
          \ id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
        collectionFormat: "multi"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "limit returning object count"
        required: false
        type: "integer"
        default: 2
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/UserPagination"
      x-swagger-router-controller: "User"
    post:
      tags:
      - "user"
      summary: "create user"
      description: "create user"
      operationId: "userPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create user"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "User"
    patch:
      tags:
      - "user"
      summary: "update user"
      description: "update user"
      operationId: "userPATCH"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create user"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "User"
  /user/{Id}:
    get:
      tags:
      - "user"
      summary: "get a user by ID"
      description: "get a user by ID"
      operationId: "userIdGET"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of user to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
      x-swagger-router-controller: "User"
    delete:
      tags:
      - "user"
      summary: "Destroy a user by ID"
      description: "delete a user by ID"
      operationId: "userIdDELETE"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of user to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "User"
definitions:
...
  RealEstateVisit:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        description: "Id de la visite"
      id_real_estate:
        type: "integer"
        format: "int64"
        description: "Id du bien"
      id_booker:
        type: "integer"
        format: "int64"
        description: "Id du booker"
      id_visitor:
        type: "integer"
        format: "int64"
        description: "Id du visiteur"
      start_date:
        type: "string"
        format: "date-time"
        description: "Date de début (YYYY-mm-dd)"
      end_date:
        type: "string"
        format: "date-time"
        description: "Date de fin (YYYY-mm-dd)"
      start_time:
        type: "string"
        format: "date-time"
        description: "Heure de début (hh:mm)"
      end_time:
        type: "string"
        format: "date-time"
        description: "Heure de fin (hh:mm)"
      booker_is_ready:
        type: "integer"
        format: "int64"
        description: "Le booker est prêt ?"
      visitor_is_ready:
        type: "integer"
        format: "int64"
        description: "Les visiteurs sont prêts ?"
    example:
      end_date: "2000-01-23T04:56:07.000+00:00"
      booker_is_ready: 3
      start_time: "2000-01-23T04:56:07.000+00:00"
      id_visitor: 9
      id_booker: 7
      end_time: "2000-01-23T04:56:07.000+00:00"
      visitor_is_ready: 2
      id: 5
      id_real_estate: 2
      start_date: "2000-01-23T04:56:07.000+00:00"
  RealEstateVisitPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
        description: "json repose code"
      total:
        type: "integer"
        format: "int32"
        description: "total numbers"
      offset:
        type: "integer"
        format: "int32"
        description: "offset"
      limit:
        type: "integer"
        format: "int32"
        description: "limit"
      list:
        type: "array"
        items:
          $ref: "#/definitions/RealEstateVisit"
    example:
      total: 6
      code: 0
      offset: 1
      limit: 5
      list:
      - end_date: "2000-01-23T04:56:07.000+00:00"
        booker_is_ready: 3
        start_time: "2000-01-23T04:56:07.000+00:00"
        id_visitor: 9
        id_booker: 7
        end_time: "2000-01-23T04:56:07.000+00:00"
        visitor_is_ready: 2
        id: 5
        id_real_estate: 2
        start_date: "2000-01-23T04:56:07.000+00:00"
      - end_date: "2000-01-23T04:56:07.000+00:00"
        booker_is_ready: 3
        start_time: "2000-01-23T04:56:07.000+00:00"
        id_visitor: 9
        id_booker: 7
        end_time: "2000-01-23T04:56:07.000+00:00"
        visitor_is_ready: 2
        id: 5
        id_real_estate: 2
        start_date: "2000-01-23T04:56:07.000+00:00"
...
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        description: "Id de l'utilisateur"
      email:
        type: "string"
        format: "string"
        description: "Email de l'utilisateur qui sert aussi de login"
      password:
        type: "string"
        format: "string"
        description: "Mot de passe de l'utilisateur"
      avatar:
        type: "string"
        format: "string"
        description: "Url de l'avatar de l'utilisateur"
      first_name:
        type: "string"
        format: "string"
        description: "Prénom de l'utilisateur"
      last_name:
        type: "string"
        format: "string"
        description: "Nom de l'utilisateur"
      address:
        type: "string"
        format: "string"
        description: "Adresse"
      zip_code:
        type: "string"
        format: "string"
        description: "Code postal"
      city:
        type: "string"
        format: "string"
        description: "Ville"
      latitude:
        type: "string"
        format: "string"
        description: "Latitude"
      longitude:
        type: "string"
        format: "string"
        description: "Longitude"
    example:
      password: "password"
      address: "address"
      city: "city"
      latitude: "latitude"
      last_name: "last_name"
      id: 0
      avatar: "avatar"
      first_name: "first_name"
      email: "email"
      zip_code: "zip_code"
      longitude: "longitude"
  UserPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
        description: "json repose code"
      total:
        type: "integer"
        format: "int32"
        description: "total numbers"
      offset:
        type: "integer"
        format: "int32"
        description: "offset"
      limit:
        type: "integer"
        format: "int32"
        description: "limit"
      list:
        type: "array"
        items:
          $ref: "#/definitions/User"
    example:
      total: 6
      code: 0
      offset: 1
      limit: 5
      list:
      - password: "password"
        address: "address"
        city: "city"
        latitude: "latitude"
        last_name: "last_name"
        id: 0
        avatar: "avatar"
        first_name: "first_name"
        email: "email"
        zip_code: "zip_code"
        longitude: "longitude"
      - password: "password"
        address: "address"
        city: "city"
        latitude: "latitude"
        last_name: "last_name"
        id: 0
        avatar: "avatar"
        first_name: "first_name"
        email: "email"
        zip_code: "zip_code"
        longitude: "longitude"
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      msg:
        type: "string"
    example:
      msg: "msg"
      code: 0
...